weblcm-python: add swupdate precheck script to add custom logic during update
This check for a script that could be placed on the system allows for addition
work to be done before an update. It could even be used to abort an update and
provide an appropriate error message for why the update was aborted.
Move all weblcm-python .sh files to /etc/weblcm-python/scripts/

bug 12537