Allow Bluetooth devices to be specified in the form "xx_xx_xx_xx_xx_xx" in addition to the already supported form "xx:xx:xx:xx:xx:xx".
MAC addresses are more typically used in the format "xx:xx:xx:xx:xx:xx", and Bluetooth device scan results return them in this format.
With the redesigned Bluetooth REST API endpoints, we specify the device in the URI (URL).  Strictly speaking, ":" is a reserved character for URIs.  This change will allow customers to specify the device target in a compliant manner, as well as support the legacy format for existing code.  Supporting the legacy format will not violate any rules, and could be left as an undocumented feature.

Bug 19642