Post error with ErrorMsg if /bluetooth controller query parameters does not contain any known values.
REST query of e.g. "https://192.168.1.108/bluetooth/controller0?filter=deadbeef" was returning success with empty results, which could be confusing for customers.  In cases where we do not recognize any of the specified filter items, we now return error with a description to inform the user that the filter value was not recognized.

Bug 19645