weblcm: Add support for "transportFilter" as a configurable Bluetooth adapter property.
Last value set wins, as with standard BlueZ API.

Setting the transport filter to "le" or "bredr" may improve scan results reliability in scenarios where the target devices type is known.

Unfortunately, read of transportFilter will only return the value we last set for the adapter in question, which may not match what's actually in effect, because according to the BlueZ docs, "When multiple clients call SetDiscoveryFilter, their filters are internally merged."  Additionally, the BlueZ method GetDiscoveryFilters surprisingly does not return the actual filters in effect on the adapter.

Bug 19905