AWM support corrections
AWM geolocation enable/disable scanning is only available when in LITE mode.
To be in LITE mode, there needs to be a -M LITE option in the
/etc/default/adaptive_ww file.  To be able to adjust the scan_attempts,
there must be a writable file in the /data/ area and this file name must
be set in the weblcm_settings.ini file.  Only when all these conditions
are met can enable/disable geolocation scanning be controlled.

BUG 17536