weblcm-python: /bluetooth: add management features for HID and VSP sockets
Add "hidList", "gattList", "hidDisconnect", and "gattDisconnect" handling such that users can query for open hid / vsp services, and close them.

This could be important if, for example, an end-user needs to change to a different remote or barcode scanner.  Previously, the service would have to be left on a (serving) port and that port abandoned, not reused.

Bug 20237