rework network connection POST endpoint.
The connection endpoint was attempting to allow update to portions of a
configuration, but was causing portions to be omitted.
The commit reworks the connection POST command to allow creation or
replacement of connections.  Any modification to exiting connection
should be merged by the user and sumbitted as a replacement connection.

BUG 20260