weblcm-python: /bluetooth: Move open/close of firewall ports out of TCP connection thread.
Opening the firewall port in the TCP connection thread results in REST API VSP/HID connect command returning prior to the service actually being available to the remote host.

By opening (and closing) the firewall ports outside of the connection thread, the REST command will not complete until the port is opened, avoiding this race condition with the host.

Bug 20371