weblcm-python: /bluetooth: Take no further action if transport filter set fails.
LCM REST API bundles state requests for the Bluetooth controller into one JSON request packet.  For example, a transport filter and the discovering state may be set by a single REST command.  In the event that setting the transport filter fails, we should return an error to the user, and not change the discovering or discoverable state.

Bug 20396