weblcm-python: bluetooth: restore VSP connections when BT radio resets occur
Track changes applied to BT controllers, devices, and VSP services, and shutdown and restore them if BT controller is disconnected from dbus / reconnected to dbus.

This handles the case when WiFi device recovery kicks in and resets both the WiFi and BT radios.  Controller reconnection to dbus correlates with HCI_DEV_UP event, but is less complex to register for.

Also tears-down VSP connections when corresponding devices are unpaired (removed).

Bug 20731