weblcm-python: Bluetooth: correct connect handling when BLE module loaded
Remove double-connect attempt when BLE module is loaded, replacing it with an improved connect in IG's BLE connect, and optionally raise connect errors so that they can be reported over the REST API as is the behavior when the BLE module is not loaded.
Also correct some syntax issues flake8 was complaining about.

Bug 21540