weblcm-python: Add new endpoint to return all NetworkManager properties for a given interface
Introduce a new endpoint (GET to /networkInterface) which returns all available
NetworkManager Device properties as defined in the D-Bus API documentation:

https://developer-old.gnome.org/NetworkManager/stable/gdbus-org.freedesktop.NetworkManager.Device.html

In addition, a complementary usage example script is included as well.

Also, the underlying logic for the /networkStatus endpoint (which is reused in
the new /networkInterface endpoint) was expanded to return extra information
(including DNS information).

Bug 22030