weblcm-python: datetime: defer generation of the zones list
On some networks, it was discovered that "timedatectl list-timezones" intermittently timed out after 10 seconds and timezone list generation failed.  Due to the fact that this appears to be a network-sensitive issue, just defer generation of the timezone list until it is requested (at present, only by REST API call which must be over a network).  Hopefully, because the request won't come in until the network is up, the timezone list will also generate in a fraction of a second then as it nominally does.  This avoids arbitrarily increasing the timeout value for particular networks.

Bug 22139