weblcm-python: Add support for controlling journal log forwarding
This change adds support for controlling and querying the active state of the
systemd journal log forwarding service (systemd-journal-gatewayd). The
service's active state is read and set via the systemd D-Bus API.

The desired behavior is that the log forwarding service is disabled by default,
but, after it is enabled, the service should be persistently activated at boot
by systemd. To achieve this, an empty 'flag file'
(/data/secret/weblcm-python/log_forwarding_enabled) is created/removed  which
can be checked by systemd as a pre-start condition (ConditionPathExists) for the
journal forwarding service.

Bug 22085