weblcm-python: Update method for reading journal logs
This change removes the use and dependency on the aging and infrequently
maintained python-systemd package (the most recent release targets systemd 234,
and we're currently using systemd 250).

The python-systemd package was being used to read the journal logs as JSON, but
this change switches to using journalctl itself with the built-in switch
'--output=json'.

Additionally, the python example script to read the journal logs which also
utilized the python-systemd package has been removed.

Lastly, a couple small formatting fixes as suggested by the linting tool Flake8
were made.

Bug 22304