weblcm-python: Add endpoint to retrieve certificate info
This change adds a new endpoint which can either retrieve pertinent metadata
about an installed certificate based on a provided cert name and password or
list the available certificates (a wrapper around /files?type=cert).

Reading the certificate metadata is delegated to the openssl_extension
Python C module which uses the OpenSSL API directly and returns the cert
info as a Python dictionary.

Bug 22040