weblcm-python: Fix logic for determining connection 'activated' state
This change addresses an issue where we were conflating a NetworkManager
profile being 'active' with an ActiveConnection being in the 'Activated' state.
In short, a profile can be 'active' (connecting, associating, etc.) without
being yet in the 'Activated' state.

A number of null-checks are added for some libnm data-retrieval functions which
can return an object or 'None'. Without these checks, we could attempt to
access the members of a 'None' object and unecessarily hit an exception.

Bug 22959