weblcm-python: Enable support for SSL client authentication and make sessions optional
This change introduces optional support for SSL client authentication via
Cheroot's (the CherryPy webserver) BuiltinSSLAdapter. This requires the use of
the OpenSSL CERT_REQUIRED 'verify mode' and results in OpenSSL rejecting a
connection from a client which doesn't present a certificate or presents an
invalid one.

One special note is that, in order to ensure proper communication on embedded
systems, there is a need to disable OpenSSL's verification of the certificate's
validity dates. Newer (OpenSSL 1.1.1+) versions of OpenSSL support the
X509_V_FLAG_NO_CHECK_TIME 'verify flag' to enable this functionality; however,
the FIPS version (1.0.2) does not. Therefore, a custom 'verify_mode'
(PY_SSL_CERT_REQUIRED_NO_CHECK_TIME) has been added to the Python3 SSL module to
patch in this logic.

Additionally, the login logic in the backend and UI has been updated to support
when cookie-based sessions are disabled.

Bug 23070