weblcm-python: Always verify credentials at login
Previously, the /login endpoint would not verify the provided credentials if a
valid session cookie was provided, and this would result in a user being able to
login with invalid credentials (with a valid cookie).

This change now always verifies the provided credentials and invalidates the
session cookie if the verification fails.

Bug 23012