weblcm-python: Add ability to import/export NetworkManager connections and certificates
This change adds new logic to the /files endpoint to support import/export of
NetworkManager connections and certificates. A GET request to the
/files?type=network endpoint will create a zip file containing the
NetworkManager connections and certificates which is encrypted with the required
password parameter. A PUT request to the /files?type=network endpoint will do
the reverse: extract an encrypted zip file using the required password parameter
to the proper locations and then trigger NetworkManager to reload the connections
from disk.

Also included in this change are a pair of usage example scripts to utilize the
new endpoint as well as various linting fixes.

Bug 22664