weblcm-python: Use Boot Time for session expiration
By default, CherryPy uses the system time to determine when a session expires;
however, this can cause clients to be unexpectedly unauthenticated when the
system time changes. This change overrides the default behavior to use
Boot Time (the time since the system was started including any time when the
system is in suspend). See below for more details on Boot Time:

https://linux.die.net/man/2/clock_gettime

Bug 24272