weblcm-python: VSP: Remove blocking on GATT writes and queue data in TCP socket
For VSP "raw" mode, this change now removes the logic to block until a GATT
write (WriteValue()) returns in order to increase throughput from the TCP host
to the GATT client.

However, if ACK'ed GATT writes are used, this can introduce 'InPrgoress' errors
(dropped data). To address this, this change also now waits to receive data
(socket.recv()) until the previous GATT write has completed.

Additionally, we now call a thread-safe callback when a VSP connection is closed
which is used to remove the connection from the internal list stored in the
instance of the VspConnectionPlugin.

Also added are a number of input validation checks before starting the VSP
connection.

Bug 24372