weblcm-python: VSP: Improve TCP client socket cleanup and ensure new clients are accepted
This change introduces the addition of return values for the socket watch event
handlers - True keeps the watch event handler alive, and False removes it.
Returning True in server_socket_event_handler() allows the socket to accept new
incoming client connections.

Additionally, it was discovered that the REST endpoint handlers are run in a
different thread than the GLib mainloop thread, so, calls into the VSP
functionality (vsp_close(), gatt_connect()) are now wrapped with
GLib.idle_add() to preserve the single-threaded nature of the VSP.

Finally, 'tcp_connection' was renamed to 'tcp_client_socket' for greater clarity
and readability.

Bug 24455